1/12

+++ devices-server/src/controllers/usersControllers.js
@@ -5,10 +5,26 @@ module.exports ={
         try {
             const decorded = request.decorded;
             if(!(decorded.role== "administrator" )){
-                reply.code(401).send(Error("You cannot access to users")); 
+                reply.code(401).send(Error("You cannot access to the users")); 
                 return;
             }
             const user = request.body;
+            if (user.email == null ){
+                reply.code(406).send(Error("Email should not be empty"));
+                return;
+            }
+            if (user.pass == null){
+                reply.code(406).send(Error("Password should not be empty"));
+                return;
+            }
+            if (user.role == null){
+                reply.code(406).send(Error("Role should not be empty"));
+                return;
+            }
+            if (!(user.role == "administrator" ||  user.role=="user")){
+                reply.code(406).send(Error("Role malformed"));
+                return;
+              }
             const newUser = await User.create(user);
             reply.code(200).send(newUser);
         }catch (e){
@@ -19,7 +35,7 @@ module.exports ={
         try {
             const decorded = request.decorded;
             if(!(decorded.role== "administrator" )){
-                reply.code(401).send(Error("You cannot access to users")); 
+                reply.code(401).send(Error("You cannot access to the users")); 
                 return;
             }
             const users = await User.find();
@@ -32,11 +48,15 @@ module.exports ={
         try {
             const decorded = request.decorded;
             if(!(decorded.role== "administrator" )){
-                reply.code(401).send(Error("You cannot access to users")); 
+                reply.code(401).send(Error("You cannot access to the user")); 
                 return;
             }
             const userId = request.params.id;
             const user = await User.findById(userId);
+            if(user == null){
+                reply.code(406).send(Error("This ID is wrong"));
+                return;
+              }
             reply.code(200).send(user);
         }catch (e){
             reply.code(500).send(e);
@@ -46,11 +66,32 @@ module.exports ={
         try {
             const decorded = request.decorded;
             if(!(decorded.role== "administrator" )){
-                reply.code(401).send(Error("You cannot access to users")); 
+                reply.code(401).send(Error("You cannot access to the users")); 
                 return;
             }
             const userId = request.params.id;
             const updates = request.body;
+            const user = await User.findById(userId);
+            if(user == null){
+              reply.code(500).send(Error("This id is wrong"));
+              return;
+            }
+            if (updates.email == null ){
+                reply.code(406).send(Error("Email should not be empty"));
+                return;
+            }
+            if (updates.pass == null){
+                reply.code(406).send(Error("Password should not be empty"));
+                return;
+            }
+            if (updates.role == null){
+                reply.code(406).send(Error("Role should not be empty"));
+                return;
+            }
+            if (!(updates.role == "administrator" || updates.role=="user")){
+                reply.code(406).send(Error("Role malformed"));
+                return;
+              }
             await User.findByIdAndUpdate(userId, updates);
             const userToUpdate = await User.findById(userId);
             reply.code(200).send(userToUpdate);
@@ -62,11 +103,15 @@ module.exports ={
         try {
             const decorded = request.decorded;
             if(!(decorded.role== "administrator" )){
-                reply.code(401).send(Error("You cannot access to users")); 
+                reply.code(401).send(Error("You cannot access to the user")); 
                 return;
             }
             const userId = request.params.id;
             const userToDelete = await User.findById(userId);
+            if (userToDelete == null){
+                reply.code(500).send(Error("This id is wrong"));
+                return;
+            }
             await User.findByIdAndDelete(userId);
             reply.code(200).send({ data : userToDelete })
         }catch (e){
@@ -77,12 +122,16 @@ module.exports ={
         try {
             const decorded = request.decorded;
             if(!(decorded.role== "administrator" )){
-                reply.code(401).send(Error("You cannot access to users")); 
+                reply.code(401).send(Error("You cannot access to the users")); 
                 return;
             }
             const userIds = request.body;
+            if (userIds.length == 0){
+                reply.code(406).send(Error("IDs should not be empty"));
+                return ;
+              }
             let userToDeletes=[];
-            for (const userId in  userIds){
+            for (const userId of  userIds){
                 const userToDelete = await User.findById(userId);
                 await User.findByIdAndDelete(userId);
                 userToDeletes.push(userToDelete);
